Sprite test cleanup
